shipsense-openai/requirements.txt
Flask~=3.0.3
SQLAlchemy~=2.0.31
openai~=1.35.7
langchain~=0.2.6
langchain_community
pymysql
elasticsearch~=8.14.0
python-dotenv~=1.0.1
fitz~=0.0.1.dev2
PyMuPDF